requirements.txt
py-cord
openai
python-dotenv
pynacl
deepgram-sdk
requests
aiohttp
moviepy